# internal/server/web/openapi.yaml
...
openapi: 3.1.0
info:
  title: wudict
  version: "1"
  summary: A local dictionary server for MDX/MDD, StarDict, Slob, DSL and BGL files.
  description: |
    WuWeiDict serves your own dictionary files over HTTP, on your own machine.

    **Streaming.** `/api/dicts`, `/api/search` and `/api/rescan` answer with
    NDJSON: one JSON object per line, flushed the moment it is ready. Read them
    line by line; do not wait for the body to end. That is the whole point - the
    first dictionary's results arrive while the slowest one is still opening.
    Every stream opens with a `begin` line and closes with an `end` line.

    `/api/ingest` streams Server-Sent Events instead, because it reports
    progress rather than results.

    **Switches.** Every boolean query parameter reads the same spellings: off
    for `0`, `false`, `no`, `off` or an empty value, on for anything else.

    **Two surfaces.** Three read-only endpoints answer a browser extension
    cross-origin: `/api/dicts`, `/api/search` and `/res/`. Everything else is
    same-origin, is marked `internal` below, and may change between releases.

    **An access key, when the server is reachable.** On the default
    `127.0.0.1` bind there is nothing to authenticate and no key is asked for.
    On any other address - and Go's wildcard listen is dual-stack, so `0.0.0.0`
    includes every globally routable IPv6 address the machine holds - `AUTH`
...
    and let the browser keep the cookie it sets. Without it: `401`, with
...

...

...
servers:
...
security: []
...
paths:
  /api/search:
    get:
      tags: [search]
      operationId: search
      summary: Search every dictionary, streaming
      description: |
        Queries the selected dictionaries concurrently and streams one line per
        dictionary as it finishes. Lines arrive in completion order, never in
        slot order: each `hit` carries `i`, its index in the `begin` line's
        `slots` array.

        A dictionary that fails, that cannot serve the requested mode, or that
        was declined by the memory cap reports that on its own line. None of
        those is a request failure, and the other dictionaries still answer.

        A search is cancelled after 30 seconds.
      parameters:
        - name: q
          in: query
          required: true
          description: The search text. Blank is a 400.
          schema: { type: string }
          example: flight
        - name: mode
          in: query
          description: |
            `exact` · `prefix` · `contains` (trigram substring over folded
            headwords) · `fts` (full text). `fuzzy` is accepted as a legacy
            alias and now means `prefix`.

            `contains` and `fts` need the dictionary to carry that index; ask
            `/api/dicts` for `caps` first, and expect `skipped` if you do not.

            A multi-word `fts` query is a PHRASE first. It is retried as a
            proximity and then as separate words only when the stricter reading
            returns nothing, and the hit says which reading answered (`rung`).
            Quotes and the operators `AND`, `OR`, `NOT` and `NEAR(a b, n)` are
            parsed when present, and an explicit query like
            `"faux ami" NOT linguistics` is answered exactly as written, without
            relaxation. Anything unparseable is read as plain words rather than
            rejected.
          schema:
            type: string
            enum: [exact, prefix, contains, fts, fuzzy]
            default: prefix
        - name: hl
          in: query
          description: |
            Mark the query's words inside each article body, wrapped in
            `<mark class="wu-hl">`. Read ONLY in `mode=fts`, where the reason a
            result matched is invisible otherwise; ignored in every other mode,
            where the match is the headword itself.

            Off unless you ask, so an article you fetch is the dictionary's own
            bytes by default. wudict's own UI sends `hl=1` from a remembered
            preference, which is why the default here is off and the default
            there is on. Marking is prefix-based and whole-token, matching what
            the full-text index matched, and is capped at 500 marks per article
            - past the cap the article is still returned complete, just no
            longer marked. `format=text` drops the marks with all other markup.
          schema:
            type: boolean
            default: false
        - name: format
          in: query
          description: |
            `raw` is the dictionary's own HTML. `clean` keeps structure,
            emphasis and media and drops scripts, styles and presentation, at
            roughly half the bytes; it also rewrites root-absolute links such
            as `/res/…` into absolute URLs, so the payload works inside a page
            served from somewhere else. `text` is markup-free, roughly 0.4.

            Ask for `clean` unless you render the article with the dictionary's
            own stylesheet.
          schema:
            type: string
            enum: [raw, clean, text]
            default: raw
        - name: dict
          in: query
          description: >
            `all`, one dictionary id, or a comma-separated list of ids in the
            order you want them back. Ids come from /api/dicts. A list where no
            id resolves is a 404; unknown ids inside a resolving list are
            skipped. Naming exactly one dictionary bypasses the memory cap,
            because a search of one is a direct demand.
          schema: { type: string, default: all }
          example: oxford,webster
        - name: 'n'
          in: query
          description: Maximum results per dictionary.
          schema: { type: integer, minimum: 1, default: 20 }
      x-codeSamples:
        - lang: Shell
          label: curl
          source: |
            curl -sN 'http://127.0.0.1:6888/api/search?q=flight&mode=prefix&format=clean&n=5'

            # -N is the whole point: without it curl buffers, and the lines that
            # were sent one at a time arrive together at the end.
      responses:
        '200':
          description: The NDJSON stream. One object per line.
          content:
            application/x-ndjson:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/SearchBegin'
                  - $ref: '#/components/schemas/SearchHit'
                  - $ref: '#/components/schemas/StreamEnd'
                discriminator: { propertyName: t }
              example: |
                {"t":"begin","i":0,"slots":[{"dict":"oxford","name":"oxford"},{"dict":"webster","name":"webster"}]}
                {"t":"hit","i":1,"dict":"webster","name":"Webster's Revised Unabridged","results":[{"Headword":"flight","Body":"<div>…</div>"}]}
                {"t":"hit","i":0,"dict":"oxford","name":"Oxford Advanced Learner's","results":[]}
                {"t":"end","i":0}
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    options:
      tags: [search]
      operationId: searchPreflight
      summary: CORS preflight
      description: >
        Answers 204 with the extension grant. The extension's own requests are
        CORS-simple and never preflight; this exists so a future one does not
        meet a 405 that explains nothing.
      responses:
        '204': { description: Preflight accepted. }
  /api/dicts:
    get:
      tags: [search]
      operationId: listDicts
      summary: List every dictionary, streaming
      description: >
        `begin` carries the total immediately, from registry ids alone, so a
        client can show "0 of 105" before any dictionary is opened. Rows follow
        in completion order.
      x-codeSamples:
        - lang: Shell
          label: curl
          source: |
            curl -sN 'http://127.0.0.1:6888/api/dicts'

            # just the ids, which is what /api/search wants
            curl -sN 'http://127.0.0.1:6888/api/dicts' | jq -r 'select(.t=="dict") | .dict.id'
      responses:
        '200':
          description: The NDJSON stream. One object per line.
          content:
            application/x-ndjson:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/DictsBegin'
                  - $ref: '#/components/schemas/DictsRow'
                  - $ref: '#/components/schemas/StreamEnd'
                discriminator: { propertyName: t }
              example: |
                {"t":"begin","total":2}
                {"t":"dict","dict":{"id":"oxford","name":"Oxford Advanced Learner's","format":"mdx","path":"/Users/me/Dicts/oald.mdx","entries":184000,"caps":{"Exact":true,"Prefix":true,"Contains":false,"FTS":false}}}
                {"t":"end"}
        '500':
          $ref: '#/components/responses/ServerError'
    options:
      tags: [search]
      operationId: dictsPreflight
      summary: CORS preflight
      responses:
        '204': { description: Preflight accepted. }
  /res/{dict}/{path}:
    get:
      tags: [search]
      operationId: getResource
      summary: One file from inside a dictionary
      description: |
        Images, stylesheets, scripts and audio, as the article references them.

        `path` is the whole remainder of the URL and may contain slashes, which
        an OpenAPI path parameter cannot express: send it unencoded, exactly as
        the article asks for it.

        Speex audio is transcoded to WAV and answered as `audio/wav`. A `.spx`
        that cannot be transcoded, or one supplied in a `res/` override folder,
        is answered raw as `audio/ogg`. Files from a `res/` folder are served
        uncached, so an edit takes effect on reload.
      parameters:
        - name: dict
          in: path
          required: true
          description: Dictionary id, from /api/dicts.
          schema: { type: string }
        - name: path
          in: path
          required: true
          description: The resource path inside the dictionary, subfolders included.
          schema: { type: string }
          example: audio/flight__gb.mp3
      x-codeSamples:
        - lang: Shell
          label: curl
          source: |
            # the path after the id is the article's own path, slashes and all
            curl -s 'http://127.0.0.1:6888/res/oxford/audio/flight__gb.mp3' -o flight.mp3
      responses:
        '200':
          description: The file, with the type the dictionary declares.
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: No such dictionary, or no such file inside it.
        '500':
          $ref: '#/components/responses/ServerError'
    options:
      tags: [search]
      operationId: resourcePreflight
      summary: CORS preflight
      parameters:
        - name: dict
          in: path
          required: true
          schema: { type: string }
        - name: path
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: Preflight accepted. }
...
  /api/clear-database:
    post:
      tags: [internal]
      operationId: clearDatabase
      x-internal: true
      summary: Clear selected dictionary indexes and obsolete prepared data
      description: >
        Requires local access or ALLOW_REMOTE_DELETE. Source dictionaries and
        their resources are preserved. Index rebuilds atomically replace the
        main index, retaining optional indexes unless selected for removal.
        Obsolete cleanup includes unused prepared folders regardless of keep
        markers, duplicates, incomplete indexes, abandoned ingest files and
        packed media. Obsolete active indexes are refreshed from their sources.
        Database dictionaries supplied as sources are never cleared.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                index: { type: boolean, default: false }
                contains: { type: boolean, default: false }
                fullText: { type: boolean, default: false }
                obsolete: { type: boolean, default: false }
      responses:
        '200':
          description: Completed; failures are reported per item.
          content:
            application/json:
              schema:
                type: object
                required: [failed]
                properties:
                  failed:
                    type: array
                    items: { type: string }
        '400':
          description: Invalid request or no option selected.
        '403':
          description: Removal permission denied.
  /api/rescan:
    post:
      tags: [internal]
      operationId: updateDictionaryIndexes
      x-internal: true
      summary: Rescan folders and apply the selected index actions
      description: >
        New dictionaries receive a main index and only the selected optional
        indexes, respecting configured DSL variants. Existing optional indexes
        are updated only when already present; deletion cancels pending creation.
        Keep preserves existing feature presence. Cleanup always removes unused
        or disabled prepared indexes and legacy packed media. Sources survive.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [new, existing]
              properties:
                new:
                  type: object
                  required: [index]
                  properties:
                    index: { type: boolean, enum: [true] }
                    contains: { type: boolean }
                    fullText: { type: boolean }
                existing:
                  type: object
                  properties:
                    index: { type: string, enum: [recreate, keep] }
                    contains: { type: string, enum: [update, delete, keep] }
                    fullText: { type: string, enum: [update, delete, keep] }
      responses:
        '200':
          description: Completed; failed lists any individual failures.
          content:
            application/json:
              schema:
                type: object
                properties:
                  failed: { type: array, items: { type: string } }
        '400':
          description: Invalid index actions.
        '403':
          description: Removal permission denied.
    get:
      tags: [internal]
      operationId: rescan
      x-internal: true
      summary: Re-scan the dictionary folders, then list
      description: >
        Reads the dictionary folders again, warms the registry, and then answers
        exactly as /api/dicts does. Removes unused prepared folders (including
        previously kept caches), abandoned indexes and packed media. Source
        files and both discovered DSL variants are preserved. Requires removal
        permission, like POST /api/clear-database.
      x-codeSamples:
        - lang: Shell
          label: curl
          source: |
            # picks up files added to the dictionary folders, then lists
            curl -sN 'http://127.0.0.1:6888/api/rescan' | head -1
      responses:
        '200':
          description: The /api/dicts stream, after the rescan.
          content:
            application/x-ndjson:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/DictsBegin'
                  - $ref: '#/components/schemas/DictsRow'
                  - $ref: '#/components/schemas/StreamEnd'
                discriminator: { propertyName: t }
        '500':
          $ref: '#/components/responses/ServerError'
  /api/howto:
    post:
      tags: [internal]
      operationId: howtoCopy
      x-internal: true
      summary: Put an editable copy of the wudict howto into the import folder
      description: >
        Copies the built-in guide (`wudict-howto.wudict.md` and its
        `wudict-howto.wudict.files/`) into the folder imports go to - the first
        dictionary folder - then rescans. The copy is listed in place of the
        built-in guide. An existing copy is never overwritten.
        With `restore=1` it instead brings the built-in guide back after the
        user removed it (`DELETE /api/library?dict=wudict-howto`, which
        records the removal so it is not written back at the next start).
      parameters:
        - name: restore
          in: query
          required: false
          schema: { type: string }
          description: Bring the removed built-in guide back instead of copying it. A switch.
      responses:
        '200':
          description: Copied; `path` is the new file.
          content:
            application/json:
              schema:
                type: object
                properties:
                  path: { type: string }
        '409':
          description: No dictionary folder is configured, or the folder already holds the guide.
        '500':
          $ref: '#/components/responses/ServerError'
  /api/ingest:
    get:
      tags: [internal]
      operationId: ingest
      x-internal: true
      summary: Prepare a dictionary, or add an index to it
      description: |
        Brings one dictionary to the feature state you ask for, streaming
        Server-Sent Events:

            event: progress   data: {"done":N,"total":M}
            event: done       data: {…a /api/dicts row…}
            event: error      data: {"error":"…"}

        Each switch is a state, not a verb: a parameter you leave out keeps
        whatever that feature is now, so toggling one thing cannot strip
        another, and a request naming none prepares an unprepared dictionary
        with its headword index only. Progress is emitted at most every 200 ms.

        The change runs as the dictionary's job on the server, not inside the
        request: a client that disconnects loses only the stream, and the work
        goes on. While a change to a dictionary runs, its `/api/dicts` row
        carries `job`, and a request for that dictionary follows the running
        change instead of starting another - what it asked for is not queued,
        and the `done` it receives carries what the dictionary now has.
      parameters:
        - name: dict
          in: query
          required: true
          description: Dictionary id.
          schema: { type: string }
        - name: contains
          in: query
          description: Trigram substring index over headwords. A switch.
          schema: { type: string }
        - name: fts
          in: query
          description: Full-text index over article text - the largest one. A switch.
          schema: { type: string }
        - name: media
          in: query
          description: Pack the dictionary's media into media.db. A switch.
          schema: { type: string }
      x-codeSamples:
        - lang: Shell
          label: curl
          source: |
            # add the substring index to one dictionary and watch it build
            curl -sN 'http://127.0.0.1:6888/api/ingest?dict=oxford&contains=1'

            # leave a switch out and it keeps its current state; this only packs media
            curl -sN 'http://127.0.0.1:6888/api/ingest?dict=oxford&media=1'
      responses:
        '200':
          description: The event stream.
          content:
            text/event-stream:
              schema: { type: string }
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/user-groups:
    get:
      operationId: getDictionaryGroups
      summary: List user groups and the automatic All Dictionaries group
      description: Private collection API. Members are current dictionary IDs, including globally disabled dictionaries. Unavailable dictionaries retain their stored membership for later rescans.
      responses:
        '200':
          description: Array of groups with id, name, ordered members (dictionary IDs), and readonly.
    post:
      operationId: createDictionaryGroup
      summary: Create an empty user dictionary group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 100 }
      responses:
        '200':
          description: Created group with empty members.
        '400':
          description: Empty, overlong or invalid name.
        '409':
          description: Duplicate or reserved name (case insensitive).
        '500':
          description: State could not be saved; no group created.
  /api/user-groups/member:
    put:
      operationId: setDictionaryGroupMember
      summary: Set one dictionary's membership without changing search settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [group, dict, member]
              properties:
                group: { type: string }
                dict: { type: string }
                member: { type: boolean }
      responses:
        '200':
          description: Membership saved.
        '400':
          description: Invalid request or attempt to edit All Dictionaries.
        '404':
          description: Group or dictionary no longer exists.
        '500':
          description: State could not be saved; membership unchanged.
...
components:
  responses:
    BadRequest:
      description: A parameter is missing or unknown.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    StreamEnd:
      type: object
      description: The last line of any NDJSON stream.
      properties:
        t: { const: end }
        i:
          type: integer
          description: Always 0 here. The search stream carries the field on every line.
      required: [t]
...